Combine timing container and dialog into one component
